+++ editor/vox/vox_scene_importer.h
@@ -17,7 +17,7 @@ public:
 	int _zn_get_preset_count() const override;
 	String _zn_get_save_extension() const override;
 	String _zn_get_resource_type() const override;
-	double _zn_get_priority() const override;
+	float _zn_get_priority() const override;
 	int _zn_get_import_order() const override;
 	void _zn_get_import_options(StdVector<zylann::godot::ImportOptionWrapper> &p_out_options, const String &p_path,
 			int p_preset_index) const override;
